공휴일 삭제

+++ place-api/src/main/java/info/myplace/api/place/repository/TagRepository.java
@@ -9,5 +9,5 @@ public interface TagRepository extends JpaRepository<Tag, Long> {
 
   List<Tag> findByNameStartsWith(String keyword);
 
-  long deleteTagById(long id);
+  int deleteTagById(long id);
 }
